+++ lib/caffeinate/dripper/subscriber.rb
@@ -10,14 +10,7 @@ module Caffeinate
       end
 
       module ClassMethods
-        # Runs the subscriber_block
-        #
-        #   OrderDripper.subscribe!
-        def subscribe!
-          subscribes_block.call
-        end
-
-        # Returns the campaign's `Caffeinate::CampaignSubscriber`
+        # Returns the Campaign's `Caffeinate::CampaignSubscriber`
         def subscriptions
           caffeinate_campaign.caffeinate_campaign_subscriptions
         end
@@ -56,33 +49,6 @@ module Caffeinate
         # @return [Caffeinate::CampaignSubscriber] the CampaignSubscriber
         def unsubscribe!(subscriber, **args)
           caffeinate_campaign.unsubscribe!(subscriber, **args)
-        end
-
-        # :nodoc:
-        def subscribes_block
-          raise(NotImplementedError, 'Define subscribes') unless @subscribes_block
-
-          @subscribes_block
-        end
-
-        # The subscriber block. Used to create `::Caffeinate::CampaignSubscribers` subscribers.
-        #
-        #   subscribes do
-        #     Cart.left_joins(:cart_items)
-        #         .includes(:user)
-        #         .where(completed_at: nil)
-        #         .where(updated_at: 1.day.ago..2.days.ago)
-        #         .having('count(cart_items.id) > 0').each do |cart|
-        #       subscribe(cart, user: cart.user)
-        #     end
-        #   end
-        #
-        # No need to worry about checking if the given subscriber being already subscribed.
-        # The `subscribe` method does that for you.
-        #
-        # Optionally, can subscribe a user manually via `Caffeinate::Campaign#subscribe`
-        def subscribes(&block)
-          @subscribes_block = block
         end
       end
     end